refactor(app): split region map rendering into extension part
Implement a focused #1571 slice by extracting CtRegionMapComponent render/tile paint logic into a dedicated part file while keeping behavior unchanged. This reduces region_map_component.dart below the 800 LOC target and preserves existing map rendering boundaries.

Refs #1571

Made-with: Cursor
